Extend the Account class to mute mailing lists
To avoid notification hell, it is sometimes desirable to have a way to
_not_ get notified about mails sent via mailing lists. This commit
introduces the logic to do so.

The tell-tale used to identify such mails is the `List-Id` header, see
https://www.ietf.org/rfc/rfc2919.txt for details.

Note: this commit does not yet introduce any user interface to specify
those senders. This is conceptually different from the logic introduced
here, and is left for a later commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>