Add support for muting senders directly from the notification
When is the most likely time when a user might want to mute
notifications for a given sender? Right: when a notification is shown,
announcing a mail arrived from that very sender.

Let's add support for that, by adding yet another action in addition to
`Reply`, `Mark Read` and `Delete`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>